Repository->Service->Controller 식의 게시글 목록, 게시글 조회 테스트 및 구현

+++ board-springboot/src/main/java/com/myboard/repository/BoardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.myboard.domain.Board;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
     List<Board> findAllByDeleteYN(Boolean deleteYN);
 }
